requirements.txt
fastapi==0.129.0
pydantic==2.12.5
pydantic_settings==2.13.1
uvicorn==0.41.0
langgraph==1.0.10
langchain-ollama==1.0.1
langchain-core==1.2.17
aiohttp==3.13.3
python-json-logger==4.0.0
arize-phoenix-client==2.1.0
arize-phoenix-otel==0.15.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-aiohttp-client==0.57b0
opentelemetry-instrumentation-logging==0.57b0
openinference-instrumentation-langchain==0.1.34
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1